+++ src/Http/AuthorizationCheckerInterface.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Phpolar\Phpolar\Http;
 
-use Phpolar\Routable\RoutableInterface;
+use Phpolar\HttpRequestProcessor\RequestProcessorInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 interface AuthorizationCheckerInterface
 {
     public function authorize(
-        RoutableInterface $routable,
+        RequestProcessorInterface $routable,
         ServerRequestInterface $request,
-    ): RoutableInterface | ResponseInterface;
+    ): RequestProcessorInterface | ResponseInterface;
 }