+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 8.0.0 (2025-06-29)
+
+### Feat
+
+- use http request processor (#447)
+
+## 7.3.0 (2025-06-29)
+
+### Feat
+
+- add support for returning http statuses (#446)
+
 ## 7.2.1 (2025-06-28)
 
 ### Fix